Bug 1737835: add an "uplift request" custom field
Adds a new custom field that stores the uplift request form.
This form is to be filled out on the tip commit of a stack to
be uplifted and will replace the similar uplift request present
in Bugzilla. The field will only present itself in the UI if the
revisions's repository is associated with the `uplift` project
tag.