Bug 1734925: Send emails to reviewer group "watchers"
There's three kinds of reviewers:
* User reviewers
* Code owner reviewers
* Group (project) reviewers

"Projects" can be watched by other users, and it's useful for
onboarding (such as if new build engineers watch #build).

So, we need to parse the list of reviewers and see if any
of them have subscribers/watchers attached, and pass that
through to the "subscribers"/"recipients" output fields.