Bug 1738776: put moz-extensions directory into a volume and mount into the test environment
Add the `moz-extensions/` directory as a Docker
volume and mount that volume into the test container.
This makes `docker-compose run test_phab` use the
code local to the working directory allowing for
faster iteration when testing.

We also unset `restart-failure` in the test environment
as this causes the container to restart when any test
fails, however the container can't restart and hangs
around.