[phabricator] only save build_unit_message for failed case or coverage
Summary:
Try to improve the revision page load latency.
Jenkins test will report back to harbormaster for every single test case run, and phabricator will save all of them to database.
Currently the table is having 1.8 billion rows, and for the revision that has a lot of test cases in CI, the page load logic will read
huge amount of data, and often we see timeouts.

After this change, we will only save for failed test case (so that dev can still view the test log), and if coverage data is provided in the unit_message, we will save as well.

Test Plan: manual apply change to staging and test

Reviewers: boyang.tian

Reviewed By: boyang.tian

Task Link: DEVX-2087

Differential Revision: https://phabricator.robinhood.com/D306789