[perf] skip file policy check if it is public or readable by all users
Summary:
Phabricator is checking files' policy if they are used in timeline or feed.

For commonly used files, like a meme in the test failure message, the file will be attached to
huge amount of the revisions, and make the policy check very slow and resource hogging.

This change adds a short-circuit evaluation: if file policy is public or all_users, skip doing the
expensive check.

Test Plan: Change in staging and test.

Reviewers: boyang.tian

Reviewed By: boyang.tian

Differential Revision: https://phabricator.robinhood.com/D306850