Bug 1754076: make `Request Uplift` interact with Bugzilla
This commit adds interactions with Bugzilla when submitting an
uplift request form on Phabricator. To do this we change the
uplift request comment action from a standard textarea populated
with remarkup for the entire request, into a form with inputs
for each question in the form. We add a custom `form` PHUIXControlType
which takes a spec with a single `questions` field to populate
the values in the form.

We change the type of `DifferentialUpliftRequestCustomField::value`
from a string to a mapping array in all usages, and only convert
the mapping to a string when modifying and accessing from internal
storage. The beta uplift questions array is changed to a mapping of
`questions => defaults` for initial form population. We add a function
`getRemarkup` which converts the field mapping values into a nicely
formatted Remarkup/Markdown string.

Since the above changes allow us to access form fields individually
as typed values, we can now make decisions based on the results of
inputs in the form. We add a code path to interact with Bugzilla
when applying transaction effects (ie saving the new result to
storage). We add a remarkup comment to Bugzilla on the relevant
bug and set the `qe-verify` flag if the user indicated manual QE
testing is required.