[phabricator] add exact responsible viewer function
Summary:
when we query revisions, there is `exact(<username>)` function that matches the exact user only, and not projects or groups the user is a member of; this is useful when we search revisions that we are directly tagged on, however this function must take in a specific username, so I can't build a query generalized for everyone

there is also a `viewer()` function, which is supposed to be a general query that can be used to build dashboards, so whoever views a dashboard used query with this function only sees their own stuff, but this includes the viewer's projects and groups and everything

this diffs adds an `exact-viewer()` function, which combines the two, so we can query things match exactly the viewer

why this change: when we discussed in our team sync how the current default phabricator "Home" dashboard is not so useful and I built my own query, @brian.myers said he wanted it to. So I figure instead of teaching people the query, I can just build a new home dashboard that's more informational. this exact-viewer() function makes the query generalized so I can just build a dashboard on top this query and make it available to everyone

Test Plan:
tested locally, and on phabricator staging

when I query "exact current viewer"
{F14482326}
I get my own stuff
{F14482342}

when I query "current viewer"
{F14482353}
I get a diff where I'm the group reviewer
{F14482375}

Reviewers: brian.myers, joel.jeske

Reviewed By: brian.myers

Subscribers: brian.myers

Differential Revision: https://phabricator.robinhood.com/D643179